opea-comps/mega-service-new/requirements.txt
opea-comps
fastapi
pip install comps